Refactor to clean group command
