+++ ckan/cli/user.py
@@ -6,16 +6,13 @@ from typing import cast
 
 import six
 import click
-import os
-import magic
 
 import ckan.logic as logic
 import ckan.model as model
 from ckan.cli import error_shout
-from ckan.common import json, config
+from ckan.common import json
 from ckan.types import Context
 from ckan.lib.helpers import helper_functions as h
-from ckan.lib.uploader import get_uploader
 
 log = logging.getLogger(__name__)
 
@@ -252,77 +249,3 @@ def list_tokens(username: str):
                 name=token[u"name"], id=token[u"id"], accessed=accessed
             )
         )
-
-
-def _get_users_with_invalid_image(mimetypes: list[str]) -> list[model.User]:
-    """ Returns a list of users containing images with mimetypes not supported.
-    """
-    users = model.User.all()
-    users_with_img = [u for u in users if u.image_url]
-    invalid = []
-    for user in users_with_img:
-        upload = get_uploader('user', old_filename=user.image_url)
-        if os.path.exists(upload.old_filepath):
-            mimetype = magic.from_file(upload.old_filepath, mime=True)
-            if mimetype not in mimetypes:
-                invalid.append(user)
-    return invalid
-
-
-@user.command("clean", short_help="Clean users containing invalid images.")
-@click.option(
-    "-f", "--force", is_flag=True, help="Do not ask for comfirmation."
-    )
-def clean(force: bool):
-    """Removes users with invalid images from the database.
-
-    Invalid images are the ones with mimetypes not defined in
-    `ckan.upload.user.mimetypes` configuration option.
-
-    This command will work only for CKAN's default Upload, other
-    extensions defining upload interfaces will need to implement its
-    own logic to retrieve and determine if an uploaded image contains
-    an invalid mimetype.
-
-    Example:
-
-      ckan user clean
-      ckan user clean --force
-
-    """
-    mimetypes = config.get_value("ckan.upload.user.mimetypes")
-    if not mimetypes:
-        click.echo("No mimetypes have been configured for user uploads.")
-        return
-
-    invalid = _get_users_with_invalid_image(mimetypes)
-
-    if not invalid:
-        click.echo("No users were found with invalid images.")
-        return
-
-    for user in invalid:
-        msg = "User {} has an invalid image: {}".format(
-            user.name, user.image_url
-        )
-        click.echo(msg)
-
-    if not force:
-        click.confirm("Delete users and its images?", abort=True)
-
-    for user in invalid:
-        upload = get_uploader('user', old_filename=user.image_url)
-        try:
-            os.remove(upload.old_filepath)
-        except Exception:
-            msg = "Cannot remove {}. User will not be deleted.".format(
-                upload.old_filepath
-            )
-            click.echo(msg)
-        else:
-            model.Session.delete(user)
-            model.Session.commit()
-            msg = "User {} with image ({}) has been deleted.".format(
-                user.name, user.image_url
-            )
-            click.echo(msg)
